# openapi/plystra.v0.0.1.yaml
components:
    schemas:
...
        ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource:
            properties:
                data:
                    $ref: '#/components/schemas/ApiOpenAPIResource'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIError:
            properties:
                audit_log_id:
                    example: audit_01hzyj6b6q8
                    type: string
                code:
                    example: VALIDATION_FAILED
                    type: string
                deny_code:
                    example: SCOPE_OUT_OF_BOUNDS
                    type: string
                details: {}
                message:
                    example: Request body is invalid JSON.
                    type: string
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
                trace_id:
                    example: trace_01hzyj6b6q8
                    type: string
            type: object
        ApiOpenAPIErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ApiOpenAPIError'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIResource'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResourceType:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIResourceType'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIPagination:
            properties:
                cursor:
                    nullable: true
                    type: string
                has_more:
                    example: false
                    type: boolean
                limit:
                    example: 50
                    type: integer
            type: object
...
        ApiOpenAPIResource:
            properties:
                created_at:
                    format: date-time
                    type: string
                deleted_at:
                    format: date-time
                    nullable: true
                    type: string
                display_name:
                    type: string
                external_id:
                    type: string
                group_id:
                    type: string
                group_path:
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                owner_member_display_name:
                    type: string
                owner_member_id:
                    type: string
                resource_type:
                    type: string
                space_id:
                    type: string
                space_name:
                    type: string
                status:
                    type: string
                updated_at:
                    format: date-time
                    type: string
                visibility:
                    type: string
            type: object
...
        ApiOpenAPIResourceType:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                display_name:
                    type: string
                id:
                    type: string
                key:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                source:
                    type: string
                status:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            type: object
...
        ApiResourceMutationRequest:
            properties:
                actor:
                    $ref: '#/components/schemas/AuthzActorContext'
                display_name:
                    nullable: true
                    type: string
                external_id:
                    nullable: true
                    type: string
                group_id:
                    nullable: true
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                owner_member_id:
                    nullable: true
                    type: string
                resource_type:
                    type: string
                space_id:
                    type: string
                status:
                    nullable: true
                    type: string
                visibility:
                    nullable: true
                    type: string
            type: object
...
        AuthzActorContext:
            properties:
                binding_id:
                    type: string
                binding_status:
                    type: string
                member_display_name:
                    type: string
                member_id:
                    type: string
                member_status:
                    type: string
                relation_type:
                    type: string
                space_id:
                    type: string
                space_name:
                    type: string
                space_status:
                    type: string
                user_email:
                    type: string
                user_id:
                    type: string
                user_member_id:
                    type: string
                user_status:
                    type: string
            type: object
...
    securitySchemes:
        ApiKeyAuth:
            description: Scoped API key for server-to-server Core calls.
            in: header
            name: X-Plystra-API-Key
            type: apiKey
        BearerAuth:
            bearerFormat: opaque Plystra access token
            description: Use the access_token returned by POST /api/v1/auth/login or POST /api/v1/auth/refresh.
            scheme: bearer
            type: http
        MetricsTokenAuth:
            description: Dedicated metrics token when METRICS_TOKEN or PLYSTRA_METRICS_TOKEN is configured.
            in: header
            name: X-Plystra-Metrics-Token
            type: apiKey
info:
    description: Self-hosted Plystra Core API. Stable surfaces cover identity, authorization, resource registry, audit, admin control-plane, and scoped API keys; plugin, template, and Data Console routes are preview metadata surfaces.
    title: Plystra Core API
    version: 0.0.1
openapi: 3.0.3
paths:
...
    /api/v1/app-modules/{module_key}/resources:
        get:
            operationId: listAppModuleResources
            parameters:
                - in: path
                  name: module_key
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResourceType'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List app module resource types
            tags:
                - App Modules
...
    /api/v1/plugins/{plugin_key}/resources:
        get:
            operationId: listPluginResources
            parameters:
                - in: path
                  name: plugin_key
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResourceType'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List plugin resource types
            tags:
                - Plugins
...
    /api/v1/resources:
        get:
            operationId: listResources
            parameters:
                - description: Filter resources by Space ID.
                  in: query
                  name: space_id
                  schema:
                    description: Filter resources by Space ID.
                    type: string
                - description: Filter resources by resource type.
                  in: query
                  name: resource_type
                  schema:
                    description: Filter resources by resource type.
                    type: string
                - in: query
                  name: limit
                  schema:
                    maximum: 200
                    minimum: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List resources
            tags:
                - Resources
        post:
            operationId: createResource
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiResourceMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create a resource
            tags:
                - Resources
    /api/v1/resources/{resource_type}/{resource_id}:
        get:
            operationId: getResourceByTypeAndID
            parameters:
                - in: path
                  name: resource_type
                  required: true
                  schema:
                    type: string
                - in: path
                  name: resource_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get a resource by type and ID
            tags:
                - Resources
...
    /api/v1/spaces/{space_id}/resources:
        get:
            operationId: listSpaceResources
            parameters:
                - in: query
                  name: resource_type
                  schema:
                    type: string
                - in: query
                  name: limit
                  schema:
                    maximum: 200
                    minimum: 1
                    type: integer
                - in: path
                  name: space_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List resources in a space
            tags:
                - Resources
        post:
            operationId: createSpaceResource
            parameters:
                - in: path
                  name: space_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiResourceMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create a resource in a space
            tags:
                - Resources
    /api/v1/spaces/{space_id}/resources/{resource_id}:
        get:
            operationId: getSpaceResource
            parameters:
                - in: path
                  name: space_id
                  required: true
                  schema:
                    type: string
                - in: path
                  name: resource_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get a resource in a space
            tags:
                - Resources
        patch:
            operationId: updateSpaceResource
            parameters:
                - in: path
                  name: space_id
                  required: true
                  schema:
                    type: string
                - in: path
                  name: resource_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiResourceMutationRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIResource'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Update a resource in a space
            tags:
                - Resources
...
servers:
    - url: http://localhost:8080
...